fix: スキルパネル 「スキル入力する」用メッセージの位置調整
