+++ lib/bright_web/live/graph_live/graphs.html.heex
@@ -25,6 +25,7 @@
       num_skills={@num_skills}
       display_skill_edit_button={false}
       me={@me}
+      flash={@flash}
     />
     <div class="mt-12 lg:mt-0 flex flex-col-reverse lg:flex-row">
       <.live_component
