+++ lib/bright_web/live/skill_panel_live/skills.html.heex
@@ -27,6 +27,7 @@
         skill_panel={@skill_panel}
         query={@query}
         me={@me}
+        flash={@flash}
       />
 
       <.help_messages_area flash={@flash} />
